+++ tests/java/TestSuite.java
@@ -32,6 +32,8 @@ public class TestSuite {
         TestRunner.testClass(SVGCanvasTest.class);
         TestRunner.testClass(StreamAssetTest.class);
         TestRunner.testClass(RuntimeEffectBuilderTest.class);
+        if (OperatingSystem.CURRENT != OperatingSystem.MACOS)
+            TestRunner.testClass(VulkanTest.class);
 
         // TestRunner.testClass(TestTest.class);
         int res = TestRunner.finishTesting();